+++ app/src/main/java/com/hadysalhab/movid/screen/main/featured/MovieGroupAdapter.kt
@@ -11,7 +11,7 @@ import com.hadysalhab.movid.screen.common.ViewFactory
 class MovieGroupAdapter(private val listener: Listener, private val viewFactory: ViewFactory) :
     ListAdapter<MovieGroup, MovieGroupViewHolder>(DIFF_CALLBACK), MovieGroupItemView.Listener {
     interface Listener {
-        fun onMovieCardClicked(movieID: Int)
+        fun onMovieCardClicked(movieID: Long)
         fun onSeeMoreClicked(movieGroupType: MovieGroupType)
     }
     companion object {
@@ -37,7 +37,7 @@ class MovieGroupAdapter(private val listener: Listener, private val viewFactory:
         holder.bind(movieGroup)
     }
 
-    override fun onMovieCardClicked(movieID: Int) {
+    override fun onMovieCardClicked(movieID: Long) {
         listener.onMovieCardClicked(movieID)
     }
