+++ app/src/main/java/com/hadysalhab/movid/screen/main/featured/MovieGroupItemViewImpl.kt
@@ -41,7 +41,7 @@ class MovieGroupItemViewImpl(
         }
     }
 
-    override fun onMovieCardClicked(movieID: Int) {
+    override fun onMovieCardClicked(movieID: Long) {
         listeners.forEach {
             it.onMovieCardClicked(movieID)
         }
